Refactor to create separate versions for mobile and desktop, that share the content layout.
